CART SDK | restructure AddressAutocomplete
