+++ src/enhancers/checkout/addressAutocompleteEnhancer/NextCommerceAutocomplete.ts
@@ -2,7 +2,8 @@ import { ApiClient } from '@/api/client';
 import { useCheckoutStore } from '@/stores/checkoutStore';
 import type { AddressAutocompleteResult } from '@/types/api';
 import { nextAnalytics, EcommerceEvents } from '@/utils/analytics/index';
-import { type AutocompleteContext, createCloseButton } from './types';
+import type { AutocompleteContext } from '../types';
+import { createCloseButton } from '../utils/create-close-button';
 
 export class NextCommerceAutocomplete {
   private ctx: AutocompleteContext;
